Summary: Increases the allowable simulation time to 250 seconds.
A stroke of bad luck occured. I just spent 3 hours tracking down why something in the simulation failed. 
Eventually I determined that it was a slightly malformed xml file which happened to be cut off at 180 seconds 
by the time manager. Had the timer been set to 179 seconds the file never would have been created and the 
results script would have declared the house bad. Had the timer been 181 seconds the file would have 
completed generation an all would be well.

This was a complicated house (three storeys, lots of windows), and so it took the very longest. So I have 
increased the timer to 250 seconds. Perhaps I should increase it more.
But only 1 house of 17,000 had the issue.