Summary: Modularizes the Results printout so that data may be combined from multiple computers.
Presently the SD and DR houses are being simulated on seperate computer (with the DR taking dramatically less 
time).

In the end all of the results must end up in one place to become a spreadsheet for use with MS Excel.
To do such a transfer, the results_all variable is being turned into xml by the XML-Dumper module (must be 
installed on all computers). This modules carefully recreates the perl data structure in a file and then can 
read it back in (XML::Simple cannot).

The idea is that a future commit will take in two results xml files (results_all), merge them appropriately, 
and then call the modularized printout data.