Summary: Adds results collation and set_name checking
It was found that it was difficult to determine what the possible set_names were. The scripts were updated so any non-acceptable or non-supplied set-name will result in a warning printout of 
the possible set_names for use. This was applied for the simulation, results, etc.

A Results_collation was added to take multiple Results...xml files and collate the results. NOTE: The houses going into the collation are expected to be unique. If they are not, the second 
merge of the same house name will overwrite the first. This script is only intended to collate results that were conducted on seperate machines or on the same machine at different times. It 
is smart such that the entire database may be split evenly across two machines (where some houses of a single group (e.g. 1-SD 5-BC) will be split across the two machines and it will collate 
correctly.