The REUM results were re-assessed and found to be close, but could be made justifiably better. This was related to the original scaling factor to get the SD and DR house type values from the totals. A uniform relationship was applied based on SHEU-03.
It was discovered that this was inappropriate as additional data existed in the REUM for use in this task and that some end-uses were not uniform.
For example: SH and DHW have a breakdown in the electronic tables by house type. This is now used as the scaling factors for the same end-uses but by energy type.

For SC SHEU-03 estimates it almost entirely belongs to SD and DR types, so the entire value was used instead of a scaling.

For the AL - floor area was used. This is because it is strongly affected by floor area and not by count of dwellings or by SH relationships (HDD had not influence in the CHREM NN AL).

So this commit gives new values for comparison purposes - although they do not stray far from the previous values they are much more justifiable in the text. 



Former-commit-id: fea2ab3c74e12490608a9007458061219f11d86a