Summary: This commit adds the functionality for selecting the most appropriate BASESIMP type based on the foundation insulation placement.
Previously, type 1 (interior insulated walls) was used for basements and type 28 (uninsulated slab) was used for crawl spaces and slab on grade.

It was realized during thesis writing that this was unintended.
This commit adds a significant amount of work which was used to properly address the BASESIMP types.

Crawl spaces and slab on grade are now specified by insulation placement on the top/bottom and whether it is full or just perimeter. There are also checks for edge placement and skirt 
insulation.

Approximately 50 different types were identified for basesments corresponding with bsmt type (concrete or wood), interior/exterior wall placement, and slab placement. Only versions used by 
the CSDDRD were implemented to save time.



Former-commit-id: 353be8400297c1f05c107c19e060a3079553e325