Summary: Adds support for heating system and DHW flues to AIM-2
Unfortunately, I had never added support for flue opening consideration in AIM-2. This was because the AIM-2 
file was completed long before the HVAC file was even envisioned.

This commit reorders the generation of the files in the Hse_Gen script so that the HVAC and DHW files come 
first. Based on the equipment type and energy source, flue sizes are independently set for these systems.

The flue sizes are then inserted into the AIM file. Test cases were run and the values are impacting the 
infiltration as one would expect. Note that this also requires an update to the ESP-r source code and HOT3000 
entries had presently hardcoded the DHW flue to be shared with the heating system.