Summary: Found an error for calculating negative thermal conductivity.
There are many basement cases where insulation may be specified but the RSI is still very low.
In such cases it is desirable to increase the thermal conductivity of the insulation to reduce the RSI.

But if the difference is greater than present RSI of the insulating layer, the new condcutivity will calculate as:
k = t / (-)

As the RSI difference approaches zero, the conductivity will be huge. This is good. But the moment it passes 0 and goes negative, the conductivity will be Huge AND 
Negative. This unfortunately means that the range change on conductivity will then set it to the minimum positive value, resulting in way too much wall insulation.

This commit checks for this case and then sets it to the max conductivity.
It is expected this will not have a significant effect on energy assessment because the basement is actually being simulated with BASESIMP. This essentially will simply 
fix an issue reporting error.



Former-commit-id: 301e36eec6dd24c9d458adc2e87ecaced11918ed