Summary: Enacts changes so that results may be reinterpretted without a simulation. Previously a simulation was required to reorder the XML file. It was found that certain corrupt XML files were getting through and causing an issue at a later point. But there was no way to get back to fix them. This commit fixes that.
XML_reporting was updated so that it checks for existance of an original xml file and if it exists it uses it as the data source to build the CHREM XML file. If it doesn't, then it copies 
the xml file generated by ESP-r to the xml.orig position and then begins the examination process. This allows us to run this subroutine again after a simulation and preserve the original 
esp-r xml results. Addition updates were applied to check for NAN values.

Core_Sim was updated to handle these changes which will result in more houses being BAD, due to the extra NAN verification in XML_reporting.

Results.pl was updated to also rebuild the XML files (taking advantage of the changes in XML_Reporting). This will allow for changes on the fly to the reporting features of the CHREM.