Summary: Addition of logic for zone heat/cool and SH/SC system effectiveness reporting.
Previously, only the secondary energy items were reported. It has been determined that the ESP-r HVAC (CETC) has some issues that seem to pop up unexpectedly and cause dramatic offset in 
consumption. This was identified by comparing the energy_balance and the secondary files. Energy balance showed the zone requirements were fine, but the SH secondary consumption was 
dramatically overestimated (by an order of magnitude).

This commit adds logic to report on the zone consumption and a comparision (COP) with the SH and SC system.
A later commit will limit this to acceptable ranges.