Summary: Updates the GHG difference to account for marginal and monthly EIF.
This commit introduces the marginal EIF and monthly consideration of electricity.
Previously we were simply calculating the reduction in GHG as the difference between the two results files.
That works for energy and quantity, but not for GHG, as the difference comes from the monthly marginal EIF as opposed to the annual average EIF.

Also - note that Results.pl and module was changed to store electricity/quantity (as opposed to energy). This is due to the EIFs being specified by quantity units (e.g. kWh) as opposed to 
energy units (e.g. GJ).