Summary: Modify NG multipliers for cook-stove and clothes-dryer.
Alan Fung pointed out in my dissertation defence that EPRI conducted a study on cook stove tops. It found that electric models were twice as efficient as natural gas models at cooking food. 
Aydinalp's NN model was based on Electric cook-stoves (see Tbl 4.2 in her thesis "electric cooking appliance").
So this commit changes the multiplier to double that of electric so as to account for NG.

It also gets rid of the NG multiplier for clothes-dryers. The OEE website says there is little difference in efficiency, although there is no reference.
But when you consider it this seems appropriate as they both exhaust and thus require airflow.