This commit add the SDHW to the CHREM model.
A new xml file is defined to hold the necessary component for the three 
types of SDHW according to Kamel Haddad paper. The Database module is 
modifed acoordingly to read this xml file. 

An Input file is added to read the user defined component and Upgrade.pm 
is modified to read these inputs. 

plant template is added to templates and control and cfg file is 
modified to add the plant. 

Control.pm is changed to accept new control functions for the SDHW. For 
time being the SDHW type 2 has 4 control functions and the others has 3.

The main writing to the text file is added to the Hse_gen_sara_TOT.pl. 
In the code some parts are relocated to fulfill the requirement for 
plant implimenation.

The eligible house criteria for SDHW is changed.