This commit add the PV system to the CHREM.
The *.spm template is added to add the special material file. 
The *.elec file is modified to include the PV_array and DC_AC inverter.
The *.cfg template modified to include *.spm and electrical result file.

The optic, material and construction database modified to include teh PV 
construction. It should be noted that the PV construction which will be 
used when we model PV as a zone is fine but the other PV construction 
needs a new optical properties. This will be added in near future.

Other scripts are modified accordingly to generate and simulate houses 
with PV.

Still the result file and dal_report has to be modified.